chore(ci): split workspace tests into modules
