+++ workspace_tests/main.rs
@@ -1,0 +1,4 @@
+#![cfg(test)]
+
+pub mod toml_utils;
+pub mod version_integrity_test;
